refactoring Landing page
